drm/amd/display: Reset link encoder assignments in atomic check
[Why]
The link encoder assignments must be "invalid" before commencing
on the reassignment operation in DC. A failing atomic check from
interruption of a signal could often leave the assignment change
in place and trigger the assertion.

[How]
As a work-around the link encoder assignments are always reset
before going into DC validation in DM atomic check.

Signed-off-by: Stylon Wang <stylon.wang@amd.com>